// frontend/src/types/auth.ts
export interface GoogleCredentialResponse {
    credential?: string;
    select_by?: string;
}

export interface GoogleJwtPayload {
    iss: string;
    nbf: number;
    aud: string;
    sub: string;
    email: string;
    email_verified: boolean;
    azp: string;
    name: string;
    picture: string;
    given_name: string;
    family_name: string;
    iat: number;
    exp: number;
    jti: string;
    hd?: string;
}